requirements.txt
streamlit
PyMuPDF
Pillow
opencv-python-headless
reportlab
streamlit-pdf-viewer
